requirements.txt
pandas
plotly
streamlit
azure-identity
msgraph-core
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client